Streamline array string scanning a bit.
Eliminates some direct calls to glyph scanner functions.  I'm trying to
reduce these in favour of `find_char<...>()`, so that I can be trivially
certain (also as the code continues to evolve) that searching for ASCII
characters is the only fundamental operation that requires knowledge of
encodings.

Once we have that certainty, we can treat all "ASCII-safe" encodings
(such as UTF-8!) as monobyte encodings.  That eliminates a bunch of
code, and replaces it with much more efficient loops.