Can't test this at compile time.  Thanks VS.  :-(
For unclear reasons there's some code here that Visual Studio can't
quite run at compile time.  It's not a problem for clang & gcc.  Out on
the internet somebody found that switching cases around in a `switch`
made the difference, but not for me.

This is a real shame because testing at compile time doesn't just make
the tests run faster (and unlike the tests, compilation usually happens
in parallel) but also, it flags any undefined behaviour.  In gcc it even
flags failed assumptions.