Test a hard SJIS decoding case.
Adds a test for the original problem that triggered the decoding
revolution in libpq: the Katakana letter "so", encoded in SJIS, ends in
a byte that looks like the ASCII backslash.

Nice to know that this passed right away.  The hard part was figuring
out what the Japanese character was.