Treat ASCII-safe encodings as monobyte.
Fixes: https://github.com/jtv/libpqxx/issues/957

This is the one I've been working towards.  It's been a lot of work, but
it feels like a well-earned victory.  Drop specific support for 6 text
encodings because we can now treat them as monobyte.

In order to get here, I had to rewrite all encoding-specific code to use
just 2 primitives:

1. `find_ascii_char()` to look for any one of some specific ASCII chars.
2. compare byte to specific ASCII value.

The latter is safe only when either the code knows it's looking at the
first byte of a character; or the encoding is ASCII-safe, meaning that
no multibyte character can contain a byte whose value lies in the ASCII
range.

With all that in place, we can treat all ASCII-safe encodings as if they
were single-byte encodings.  Which is going to speed up some loops,
reduce binary sizes, and drop some finicky code that I will no longer
need to maintain.