Fix `clang-tidy` warning about null pointer.
This code assumed that if anybody ever decides to iterate beyond the end
of an iteration, whatever happens is your own problem.  But `clang-tidy`
didn't care much for that, so I gave it some predictable behaviour.

I hope performance costs will be minimal — I marked the most likely code
path as `[[likely]]`.