Forbid impossible conversions to borrowed ranges.
Some functions store data in a string with automatic duration, convert
its contents to a caller-supplied type, and then either return that
value or store it externally.

In those situations, use concepts to forbid conversion to types that
represent "borrowed" ranges.  Otherwise, the caller might for example
ask a function to return a `string_view` referring to a buffer that is
no longer alive after the function returns.

Also, add `final` to a bunch of classes, and add a new function
`result::opt_row_ref()` similar to `result::opt_row()`.