Use format library if float charconv not available.
This is likely to be faster and cleaner than falling back on the streams
library.

It'd probably make sense to do the same for conversion in the other
direction, at which point we could also drop the `thread_local` support
check.  But I'd need to find decent documentation first, and have some
idea of the status of compiler support.