Don't require `to_buf()`/`into_buf()` to zero-terminate.
This is a huge change, and at the same time a very small one.  In the
7.x string conversion API, the `to_buf()` and `into_buf()` functions
were expected to add a trailing zero behind the string, and `to_buf()`
functions returned `pqxx::zview`.

Now, in the 8.0 API they are no longer required to zero-terminate, and
`to_buf()` can return just `std::string_view`.

I'm making this change because a failure to zero-terminate had gone
unnoticed for some time, and I realise that this could happen because
we never actually rely on the zero-termination.  It was intended to
allow more use of C functions and such, but in practice I don't think
that's very useful any more.  It ended up being just a "belt and
suspenders" waste of bytes and cycles.

In this first step I'm not going out of my way to get rid of the
zero termination.  At any rate it'd get complicated in a lot of
situations where I still use the 7.x API.  So I'm introducing the API
change first, in the 8.x API which is not yet in production use.  Next
we can migrate more of the conversions to use the 8.x API, and as we do
that we'll be able to get rid of the terminating zeroes in more places.