Drop `into_buf()` from conversion API.  Drop zero.
The string conversions will no longer add terminating zeroes.  There's
now a single, global `into_buf()` but it calls the string traits'
`to_buf()`, not its `into_buf()`, and `memmove()`s the result into
place (which will probably be a no-op).

Not fully working yet.  This will need a lot of inspection, testing, and
static analysis.  But I love the thought of simplifying that API a bit.