Drop `converts_{to,from}_string`.
Simplifying the conversion API yet further: there's really no point in
the member constants `converts_to_string` and `converts_from_string`.
If either is not supported, just leave it out of the traits struct and
we can just catch that (if needed) using an `if constexpr` on a
`requires` expression.