Work around exasperating leak-sanitizer error.
This might not have taken me so long had I been feeling better...  We
were seeing a single, 98-byte memory leak reported by gcc 15's leak
sanitizer.  Just one, allocated inside a `std::string` for a demangled
type name in `pipeline.cxx`, even when `main()` doesn't do anything,
and only when building libpqxx as a shared library and in audit mode.

This was showing up in Arch Linux, but would probably have hit Debian
and Ubuntu eventually.  It also happens in Debian unstable.

The offending type turned out to be `std::shared_ptr<std::string>`.
Finding this out (through printf debugging) led me to the source: I
passed a `std::shared_ptr<std::string>` to `separated_list()`, which
performs string conversion, which in turn calls `name_type()` in case
it needs to compose an error message about the type.

_Dereferencing_ the `shared_ptr` I feed into `separated_list()` replaced
it with a simple `std::string`, and we have a specialisation of that
type's name built in.  That seems to get around the problem.