Drop `string_traits::into_buf()` from string conversion API. (#1028)
A bunch of simplifications:

* Drop `into_buf()` from the  conversion API.
* Drop the terminating zero from `into_buf()`.
* So, return `std::string_view`, not `pqxx::zview`.
* Drop `converts_{to|from}_string` from `string_traits`.
* A `to_buf()` for a string type can just return its input value.

Writing a string conversion gets much easier.   The global `pqxx::to_string()` just uses the traits' `into_buf()` (not `to_buf()`) and then moves the output into place in the buffer (which will often be a no-op).  Globally, not as repeated code in every conversion.

One _new_ complication though is that the result of a `to_buf()` call is only guarantee to remain valid while both the buffer _and_ the input value remain valid and in the same place in memory.  But in normal usage you'll typically read the string once and you're either done with it or you'll store it somewhere else anyway.

Reviews by Copilot helped find a bunch of small mistakes, such as typos in comments.