New helper script: `idock.sh`, wraps `docker`.
I'm sick and tired.  That's because I've been sick and it's had that
physical effect on me, but _also_ because I find myself running docker
all the time and I have to write the command line without mistakes.

This makes it easier.  A little wrapper script that runs the usual
`docker` command line, bind-mounting the libpqxx source tree.  And in
such a place that I'm still reminded to copy the directory to somewnhere
else before working on it.  I don't want to be building directly inside
my source tree on the host.