Make `no_null` the default.
This is a tradeoff.  Giving this a default creates ways of getting it
wrong.  For example, you might define a `nullness` for your type but
forget to include it in some of the code paths.

Still, I think that would become obvious pretty quickly.  Let's play
with this a bit and see what happens.