More checks (#1068)
Enable more run-time stdlib checks, as well as Valgrind.

Turns out Valgrind does not support `long double`!  And this caused one test to fail, _only_ with `gcc -O0` (not with clang, and not with optimisation).  As a workaround, I added a macro `PQXX_VALGRIND` which skips all `long double` tests.  (They're almost all generic anyway, so there isn't a huge amount of value to those tests, and of course we do run them elsewhere.)