Force inlining of some functions.
This means there will be no out-of-line definitions at all.  The reason
for doing this was that repeated copies were messing up the coverage
reports, and I didn't feel like moving the functions to a non-anonymous
namespace.

It's probably only going to be a partial solution to the coverage
problem, but I see some fringe benefits: smaller object files, slightly
faster builds (probably, and I'm expecting not by much), and no
accidental calls to unwanted out-of-line instances.