Tighten some buffer constraints.
Start on requiring that glyph scanners only be used within the text,
not past the end.

Assert that a `zview` ends in a terminating zero.  This failed in a few
places, but since the call sites worked regardless, it shows us that the
terminating zero wasn't essential.  So, replace those uses of `zview`
with `string_view`.