JOHAB support was always broken!
The encoding validation was buggy in dealing with 2-byte characters:
instead of checking the second byte, it would use the value of the first
byte again.  Bad text would not confuse the validation in any meaningful
way (so it's not an injection vector AFAICT), but it would trigger
unjustified exceptions.  So, probably nobody was using it.