Optimise UHC scanning for usual case.
The Korean UHC encoding has a special property: "trail bytes" (i.e.
bytes in multibyte characters other than its initial byte) are either
above the ASCII range, or they're in the range of the ASCII letters
(A—Z and a—z).

That makes UHC "ASCII-safe" in the special case where none of the
ASCII characters you're looking for is a letter.  And apart from tests,
for us, none of the ASCII characters we're looking for ever are!  This
will simplify the UHC search loops to the simplicity and speed of the
ASCII-safe encodings.