Merge BIG5 and UHC encoding groups.
These encodings are very simple in their essence: a "low" byte (0x7f or
less) is a single-byte character, and a "high" byte (0x80 or more) is
the first of a 2-byte character.

There's no real need to check for invalid characters.  All we need to
know is which bytes are stand-alone ASCII and which are part of
multi-byte characters.

One sad thing is that this takes away a short-lived optimisation for UHC
where we can treat it as ASCII-safe so long as we're not searching for
any ASCII _letters._  Perhaps at some point we'll just treat UHC as
ASCII-safe and stipulate that we can't do searches for ASCII letters.
Or, we just live with the missed opportunity.