Merge BIG5 and UHC encoding groups. (#1096)
Merge BIG5 and UHC encoding groups.

These encodings are very simple in their essence: a "low" byte (0x7f or
less) is a single-byte character, and a "high" byte (0x80 or more) is
the first of a 2-byte character.

There's no real need to check for invalid characters.  All we need to
know is which bytes are stand-alone ASCII and which are part of
multi-byte characters.

Since UHC is effectively ASCII-safe _for our purposes,_ treat it as
such.  We can no longer make this decision at the time where we
choose a character finder, because by that time the actual encoding
name is no longer known, just the encoding _group._  We make up
for that with some static assertions against unsafe usage.