Get rid of JOHAB support.  It was always broken. (#1097)
Get rid of JOHAB support.  It was always broken.

It seems that PostgreSQL's support for this encoding has always been
seriously incomplete.  Even if it wasn't, there are multiple variants of
this encoding, so there'd still be the question of _which variant_ would
be supported.

This was actually revealed by libpqxx unit testing; the PostgreSQL team
had never heard any complaints about it.  It's a client-only encoding,
meaning that there are no legacy databases encoded in JOHAB.  So, the
consensus right now seems to be that it's best just to stop listing it
as supported.