MSVC 2022 multidim support seems to be broken.
This compiler version _claims_ to support multidimensional indexing
operators (`operator[]`), but support doesn't actually seem to work.
It accepts the definition of a multidim `operator[]`, but then when it
sees a call to that operator, it mistakes the comman in the arguments
for a sequence operator (it warns that the first argument will be
ignored) and then crashes.

(The problem happens when compiling C++23, not when compiling C++20).

So, let's try tightening the configure check: it should notjust look at
the `__cpp_multidimensional_subscript` feature flag, but also attempt to
compile an actual call to a multidimensional index operator.