Pass conversion context when writing composites to buf.
One function here was still making up its own context, with an unknown
encoding group.  I didn't think that mattered since it's converting _to_
strings, but it needs to escape special characters — and for that, it
needs to know whether it's dealing with a real special character or with
a mere trail byte.