Parse SQL array directly into C++ container. (#1101)
This will make life a lot easier on people writing conversions
involving basic, single-dimensional SQL arrays.  You can now read the
text representation of an SQL array directly as a `std::vector` or other
standard container type.

This is not perfect.  It relies on the pair-of-iterators constructor for
your container type, and it doesn't let you pass an allocator at run
time.  But I think it's pretty useful!