Improve composites support. (#1110)
Make `composite_into_buf()` return a `zview` instead of `size_t`.  The
terminating zero is a convenience for the case where the composite is
being passed as an SQL statement parameter.  But it also simplifies
calling code because it no longer needs to construct its own view.
Which is what it was most likely to do as the very next step.

Adds a test for the case where there's a `std::vector<std::string>` in
the composite (as in #1108).  It reproduces part of the problem.

(Oh and yes, Copilot, the formatter removed a space in the new config
check.  Do I really have to mention such trivia in the description?)