Give `params` an optional encoding argument. (#1112)
Give `params` an optional encoding argument.

There's no really nice way to do this.  How do you add an optional
parameter to a constructor that takes variadic arguments?  Concepts
don't help.  There's a trick, but it just complicates things too much.

So instead I just use `if constexpr`.  To keep things convenient (and
terse, which really matters for an offhand little class like this) I
let the caller pass either an `encoding_group`, or a `connection`, or a
transaction.  Luckily this set of types is disjunct from the types that
you may want to pass as an actual parameter.  It's messy, but I think
reasonably effective.

It does complicate the explanations a bit: pass this as your first
argument _if needed._  Which you find out from an exception.  Not ideal.