More source locations (#1113)
Pass `std::source_location` in more places.

The purpose of all this is to make any error messages that we generate refer to either (by default) the location where the application called libpqxx that led to the error; or one that the caller passed in explicitly.

Unfortunately it's hard to avoid situations where the error message will point to somewhere inside libpqxx instead.